Add map extractor
